refactor api call

+++ lib/models/single_hospital_properties_model.dart
@@ -22,10 +22,10 @@ class SingleHospitalPropertiesModel {
   String? lga;
 
   @JsonKey(name: "lon")
-  int? lon;
+  double? lon;
 
   @JsonKey(name: "lat")
-  int? lat;
+  double? lat;
 
   SingleHospitalPropertiesModel();
